% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFun.R
\name{targetGrid}
\alias{targetGrid}
\title{Function to calculate a stars grid based on the x,y,z attributes that represents the coordinates of the map tile}
\usage{
targetGrid(x, y, z, tileSize = 256L)
}
\arguments{
\item{x}{The location of the tile in the x dimension}

\item{y}{The location of the tile in the y dimension}

\item{z}{The zoom level}

\item{tileSize}{The size of the tile generally 256 pixels}
}
\value{
A stars grid with all values being zero with the dimensions matching those required for the tile specified
}
\description{
Function to calculate a stars grid based on the x,y,z attributes that represents the coordinates of the map tile
}
\details{
This function is mostly useful for testing purpose.
}
\examples{
targetGrid(4, 2, 4)
targetGrid(4, 2, 4, 128)
}
